\name{mmed-package}
\alias{mmed-package}
\alias{mmed}
\docType{package}
\title{
Mediation Analysis Procedures for One or Two Mediators with Missing Mediator Data
\packageTitle{mmed}
}
\description{
A package providing methods to conduct mediation analysis and get estimates
of the effects when mediator data may be missing. Additionally, confidence 
interval methods for the mediation effect are included. 
\packageDescription{mmed}
}
\details{
Functions rquire independent and response variables, as well as matricies of 
covariates, and information about the mediators, and return estimates of the 
effects for the terms in the model. 

This package is still under development, with work being done to clarify values 
in the output from the functions, as well as more details in the help files.


\author{
John Kidd

Maintainer: John Kidd <jkidd@uvu.edu>
}
\references{
Kidd, John, et al. \emph{Mediation analysis of multiple mediators with incomplete omics data.} Genetic Epidemiology (2022).
\keyword{ mediation, indirect effect, multiple mediators, em algorithm }
\seealso{

}
\examples{
  \dontrun{
    }
}
